[Feat] #11: 컨트롤러 생성
