+++ src/main/java/com/kuit/findyou/domain/report/model/Neutering.java
@@ -1,0 +1,5 @@
+package com.kuit.findyou.domain.report.model;
+
+public enum Neutering {
+    Y, N, U
+}
