+++ src/main/java/com/kuit/findyou/domain/report/model/Sex.java
@@ -1,0 +1,5 @@
+package com.kuit.findyou.domain.report.model;
+
+public enum Sex {
+    M, F, Q
+}
