--- src/main/java/com/kuit/findyou/report/model/Neutering.java
@@ -1,5 +1,0 @@
-package com.kuit.findyou.report.model;
-
-public enum Neutering {
-    Y, N, U
-}
