--- src/main/java/com/kuit/findyou/report/model/Sex.java
@@ -1,5 +1,0 @@
-package com.kuit.findyou.report.model;
-
-public enum Sex {
-    M, F, Q
-}
