--- src/main/java/com/kuit/findyou/report/repository/AnimalFeatureRepository.java
@@ -1,9 +1,0 @@
-package com.kuit.findyou.report.repository;
-
-import com.kuit.findyou.domain.report.model.AnimalFeature;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnimalFeatureRepository extends JpaRepository<AnimalFeature, Long> {
-}
